chore: rename SDK to @inco/lightning-js, pin rc-11, drop lightning-preview
Follow the monorepo SDK rename @inco/js -> @inco/lightning-js: update all
imports (subpaths unchanged), the root overrides + backend/react workspace deps,
and the Tests workflow jq override. Pin @inco/lightning-js + @inco/lightning to
1.0.0-rc-11 and DOCKER_IMAGE_TAG to v1.0.0-rc-11. Drop @inco/lightning-preview
(unused by any contract) from overrides, contracts deps, and remappings.

bun.lock regeneration and CI validation are pending the 1.0.0-rc-11 publish.
