+++ .env
@@ -1,3 +1,3 @@
 # The version of the Inco Lightning test docker images to use
-DOCKER_IMAGE_TAG=v1.0.0-rc-8
+DOCKER_IMAGE_TAG=v1.0.0-rc-11
 PEPPER=mainnet
