+++ backend/package.json
@@ -10,7 +10,7 @@
     "generate:abis": "bun wagmi generate"
   },
   "dependencies": {
-    "@inco/js": "workspace:*",
+    "@inco/lightning-js": "workspace:*",
     "viem": "^2.28.0",
     "webpack": "^5.102.1"
   },
