+++ backend/src/test/react-example/package.json
@@ -8,7 +8,7 @@
     "test": "bun vitest run test/Inco.spec.tsx"
   },
   "dependencies": {
-    "@inco/js": "workspace:*",
+    "@inco/lightning-js": "workspace:*",
     "react": "latest",
     "react-dom": "latest",
     "viem": "^2.28.0"
