+++ contracts/package.json
@@ -3,7 +3,6 @@
   "private": true,
   "dependencies": {
     "@inco/lightning": "workspace:*",
-    "@inco/lightning-preview": "0.7.10",
     "@openzeppelin/contracts": "^5.3.0",
     "ds-test": "https://github.com/dapphub/ds-test",
     "forge-std": "https://github.com/foundry-rs/forge-std"
