+++ package.json
@@ -32,9 +32,8 @@
     "typescript": "^5.8.3"
   },
   "overrides": {
-    "@inco/lightning": "1.0.0-rc-8",
-    "@inco/lightning-preview": "1.0.0-rc-8",
-    "@inco/js": "1.0.0-rc-8",
+    "@inco/lightning": "1.0.0-rc-11",
+    "@inco/lightning-js": "1.0.0-rc-11",
     "chalk": "^4.1.2",
     "is-core-module": "2.13.1",
     "error-ex": "1.3.2",
